SDK-1703: Separate objective from filter
